+++ public/index.html
@@ -17,7 +17,10 @@
   <div class="root" id="root"></div>
 
   <script src="./modules/ajax.js"></script>
+  <script src="./modules/inputDataValidation.js"></script>
   <script src="./modules/createElements.js"></script>
+  <script src="./modules/actionScript.js"></script>
+  <script src="./modules/ajaxRequests.js" type="module"></script>
   <script src="./js/main.js" type="module"></script>
