+++ public/js/script.js
@@ -1,14 +1,4 @@
-const signIn = document.getElementById('sign-in');
-const signUp = document.getElementById('sign-up');
-
 setTimeout(() => {
   container.classList.add('sign-in');
 }, 200);
 
-const toggle = () => {
-  container.classList.toggle('sign-in');
-  container.classList.toggle('sign-up');
-};
-
-signIn.addEventListener('click', toggle);
-signUp.addEventListener('click', toggle);
