+++ public/modules/actionScript.js
@@ -1,0 +1,24 @@
+(function() {
+
+    // ----------------------------------------
+    function authorizViewAction(parent) {
+        const signIn = document.getElementById('sign-in');
+        const signUp = document.getElementById('sign-up');
+
+        setTimeout(() => {
+            container.classList.add('sign-in');
+        }, 200);
+
+        const toggle = () => {
+            parent.classList.toggle('sign-in');
+            parent.classList.toggle('sign-up');
+        };
+
+        signIn.addEventListener('click', toggle);
+        signUp.addEventListener('click', toggle);
+    }
+    window.authorizViewAction = authorizViewAction;
+
+
+}()
+);
